requirements.txt
requests==2.28.0
python-chess==1.999
chess.com==2.1.3
stockfish==3.28.0
